Clarify hash-table-pop! behavior when the table is empty
There's no mention what will happen when the table is empty,
which implies it is an error (or 'undefined behavior').  But it's
a bit ambiguous from the current document and reference
implementation---the latter is written in the way that it returns
a single unspecified value when the table is empty.

The suggested change clarify that such operation is an error,
and update the reference implementation to detect the case.